requirements-dev.txt
# Development and test dependencies
pytest==8.2.2
pytest-cov==5.0.0
httpx==0.27.0
PyYAML==6.0.1
